Finalize integrating CI
Those changes are all made to run tests in CI. The tests currently fail
and corrections of the failures will be made in subsequent commits. But
his whole setup will make sure that we check syntax, lint, validity of
ruby code, age of REFERENCE.md file (which does not exist yet, but once
it gets created the tests will make sure that it is always kept up to
date with API changes) and finally runs the puppet-rpsec tests.

Travis-CI is not accessible freely anymore, so we need to move to
something else and since github has "workflows" that's well integrated
in all of the web interface we'll go wit that.

Most of the changes in config are me copying how my other puppet modules
have been integrated into CI, which was done by mostly copying what
voxpupuli is doing.

The Rakefile was cut down so that we'll now use the rake tasks that are
already provided by other gems like puppet-strings and voxpupuli-test.
The use of puppet_metadata helps to remove duplicate management of the
test matrix: what we test is what's specified as supported in
metadata.json.