Don't run rubocop on pkg directory
This directory is created when packaging for forge and it contains
older copies of the module that we're not interested in. This exclusion
is only useful for devs of this module.

There were not justifications for excluding .erb files so let's remove
the exclusion and let rubocop do its thing.