comment out rspec-puppet-facts for now
the gems is not found for some reason when running rspec and I don't
know why.

The spec tests currently don't use the presets from rspec-puppet-facts
so I can just take it out for now. It would be nicer to use them in the
future but... we'll get there eventually.

for now at least I want to debug the tests as they are.