Fix all the more obvious spec failures
Well... time has passed and some of the test code just doesn't work
anymore in recent versions of ruby. This fixes most of the problems
simply by adapting the syntax used.

I've decided to disable the ImplicitSubject cop for now since it was
generating a bunch of messages and when I tried to remove the problem
mentioned I just ended up breaking the test code.

The errors that are still left in classes/smartd_spec are linked to
checking types and I would prefer to modify the module to use actual
data types for arguments, and thus simply just remove those tests since
they won't be useful anymore.

There's one test that's really broken and that I'll need to debug.