Use data types instead of ensure_* functions
Puppet will now ensure that the input values in params are of the
correct type so the module itself doesn't need to check for it anymore.
This also means that unit tests for bad values can be removed.

The optional parameters for *_options now only use the string type. No
need for the special value of undef.

I'm not sure why but suddenly some of the tests that were checking for
the presence of the smartmontools package in the catalog broke because
the value "present" to ensure gets renamed to "installed". The only
difference from before is the version of stdlib...