Force values of facts for devices
Currently the test adds facts blockdevice_sd* with the desired value for
the test, but in some cases some of the facts might already exist (e.g.
if the system has sda/sdb drives). In those cases the test fails because
the fact values were not overridden properly. This is currently
happening on github's CI.

To remove this case, we need to first add the facts to ensure that they
are there and then force a value override by mocking the value received
from the fact.